renamings and consolidation of formatting nodes in orig question retrieval
