--- backend/onyx/agents/agent_search/deep_search_a/initial/retrieve_orig_question_documents/nodes/format_raw_search_results.py
@@ -1,18 +1,0 @@
-from onyx.agents.agent_search.deep_search_a.initial.retrieve_orig_question_documents.states import (
-    BaseRawSearchOutput,
-)
-from onyx.agents.agent_search.deep_search_a.shared.expanded_retrieval.states import (
-    ExpandedRetrievalOutput,
-)
-from onyx.utils.logger import setup_logger
-
-logger = setup_logger()
-
-
-def format_raw_search_results(state: ExpandedRetrievalOutput) -> BaseRawSearchOutput:
-    logger.debug("format_raw_search_results")
-    return BaseRawSearchOutput(
-        base_expanded_retrieval_result=state.expanded_retrieval_result,
-        # base_retrieval_results=[state.expanded_retrieval_result],
-        # base_search_documents=[],
-    )
